Modulize common methods.
