+++ lib/cureutils/cli.rb
@@ -1,5 +1,6 @@
 # coding: utf-8
 require 'cureutils/version'
+require 'cureutils/common'
 require 'cureutils/cure_janken_manager'
 require 'cureutils/cure_date_manager'
 require 'cureutils/cure_grep_manager'
@@ -61,15 +62,14 @@ module Cureutils
                             type: :boolean,
                             desc: 'Print only the matched parts.'
     def grep(default_pat = '[:precure_name:]', filename = nil)
-      pat = default_pat.clone
-      pat = pregex2regex(default_pat) unless options['extended-regexp'.to_sym]
       manager = CureGrepManager.new
       manager.source_input(filename)
+      manager.pattern(default_pat.clone, options['extended-regexp'.to_sym])
       # Check the file discriptor to check the pipe exists or not.
       manager.option_colorize($stdout.isatty)
       manager.option_only(options['only-matching'.to_sym])
       # Print matched lines.
-      exit_status = manager.print_results(/#{pat}/)
+      exit_status = manager.print_results
       exit(exit_status)
     end
 
@@ -118,42 +118,6 @@ module Cureutils
     def janken
       manager = CureJankenManager.new
       exit(manager.janken.to_i)
-    end
-
-    private
-
-    # TODO: Common method
-    # Convert string to precure regular expression
-    def pregex2regex(regex, br_flg = false)
-      this_regex = regex.dup
-      br_ex = br_flg ? '' : '?:'
-      %w(girl_name human_name precure_name cast_name color).each do |attr|
-        expression = '\[:' + attr + ':\]'
-        precures_ex = cure_list(attr.to_sym).join('|')
-        replaced = "(#{br_ex}#{precures_ex})"
-        this_regex.gsub!(/#{expression}/, replaced)
-      end
-      this_regex
-    end
-
-    # TODO: Common method
-    def cure_list(sym)
-      list = Precure.all_stars.map(&sym)
-      list << Cure.echo[sym]
-      # Regulate cure princes human name
-      list.map do |str|
-        str.gsub!(/\(.+?\)/, '')
-        str
-      end
-      list
-    end
-
-    # TODO: Common method
-    def cure_table(to_sym, from_sym)
-      to_arr = cure_list(to_sym)
-      from_arr = cure_list(from_sym)
-      hash = Hash[[to_arr, from_arr].transpose]
-      hash
     end
   end
 end
