+++ lib/cureutils/cure_translate_manager.rb
@@ -25,44 +25,13 @@ class CureTranslateManager
     str.gsub(/\[:(.*):\]/, '\1')
   end
 
-  # Convert string to precure regular expression
-  def pregex2regex(regex, br_flg = false)
-    this_regex = regex.dup
-    br_ex = br_flg ? '' : '?:'
-    %w(girl_name human_name precure_name cast_name color).each do |attr|
-      expression = '\[:' + attr + ':\]'
-      precures_ex = cure_list(attr.to_sym).join('|')
-      replaced = "(#{br_ex}#{precures_ex})"
-      this_regex.gsub!(/#{expression}/, replaced)
-    end
-    this_regex
-  end
-
-  def cure_list(sym)
-    list = Precure.all_stars.map(&sym)
-    list << Cure.echo[sym]
-    # Regulate cure princes human name
-    list.map do |str|
-      str.gsub!(/\(.+?\)/, '')
-      str
-    end
-    list
-  end
-
-  def cure_table(to_sym, from_sym)
-    to_arr = cure_list(to_sym)
-    from_arr = cure_list(from_sym)
-    hash = Hash[[to_arr, from_arr].transpose]
-    hash
-  end
-
   def print_results
     # Create precure regular expression
     regex_tag = str2pregex(@translate_from.to_s)
     # Get patterns
-    cure_pattern = pregex2regex(regex_tag, true)
+    cure_pattern = Common.pregex2regex(regex_tag, true)
     # Get Key-Value
-    table = cure_table(@translate_from, @translate_to)
+    table = Common.cure_table(@translate_from, @translate_to)
     @in.each do |line|
       updated_line = line.dup
       line.scan(/#{cure_pattern}/).each do |pat|
